@@ -2,7 +2,7 @@ package com.github.crazyorr.ffmpegrecorder;
 
 import android.media.MediaPlayer;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.MediaController;
 import android.widget.TextView;
 import android.widget.VideoView;
